+++ packages/nextjs/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.6](https://github.com/mi6/ic-ui-kit/compare/@ukic/nextjs@3.2.5...@ukic/nextjs@3.2.6) (2025-08-06)
+
+**Note:** Version bump only for package @ukic/nextjs
+
 ## [3.2.5](https://github.com/mi6/ic-ui-kit/compare/@ukic/nextjs@3.2.4...@ukic/nextjs@3.2.5) (2025-07-24)
 
 **Note:** Version bump only for package @ukic/nextjs